+++ src/certsf/methods/__init__.py
@@ -1,5 +1,5 @@
 """Custom certified methods."""
 
-from .stirling import stirling_loggamma
+from .stirling import stirling_loggamma, stirling_loggamma_shifted
 
-__all__ = ["stirling_loggamma"]
+__all__ = ["stirling_loggamma", "stirling_loggamma_shifted"]